requirements.txt
GitPython==3.1.43
unidiff==0.7.5
networkx==3.3
fastapi==0.115.0
uvicorn==0.30.6
aiofiles==23.2.1
